+++ src/main/java/contest/collectingbox/module/location/domain/repository/DongInfoRepository.java
@@ -1,5 +1,6 @@
-package contest.collectingbox.module.location.domain;
+package contest.collectingbox.module.location.domain.repository;
 
+import contest.collectingbox.module.location.domain.DongInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
