+++ src/main/java/contest/collectingbox/module/location/domain/repository/LocationRepository.java
@@ -1,5 +1,6 @@
-package contest.collectingbox.module.location.domain;
+package contest.collectingbox.module.location.domain.repository;
 
+import contest.collectingbox.module.location.domain.Location;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LocationRepository extends JpaRepository<Location, Long> {
